requirements.txt
Flask==3.0.2
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==8.0.0
black==24.1.1
flake8==7.0.0